[OP-815] Skip stateful tests on-prem via buildinfo instead of build tags
Keep the `tf_acc_onprem_secure` build tag on the stateful rule tests and
skip them at runtime with `buildinfo.OnpremSecure`, matching the pattern
already used in resource_sysdig_secure_custom_policy_test.go and
resource_sysdig_secure_team_test.go. The tests stay compiled and report a
visible SKIP on-prem instead of disappearing from the build.

Fold the stateful step of TestAccManagedPolicyDataSource back into its
original file, gated by the same flag, and drop the separate
data_source_sysdig_secure_managed_stateful_policy_test.go: it named a
data source that does not exist, lost the IBM guard, and added an extra
provider cycle plus a job to the per-file CI matrix.
