+++ sysdig/data_source_sysdig_secure_rule_stateful_count_test.go
@@ -1,4 +1,4 @@
-//go:build tf_acc_sysdig || tf_acc_sysdig_secure || tf_acc_policies
+//go:build tf_acc_sysdig || tf_acc_sysdig_secure || tf_acc_policies || tf_acc_onprem_secure
 
 package sysdig_test
 
@@ -13,10 +13,14 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
+	"github.com/draios/terraform-provider-sysdig/buildinfo"
 	"github.com/draios/terraform-provider-sysdig/sysdig"
 )
 
 func TestAccRuleStatefulCountDataSource(t *testing.T) {
+	if buildinfo.OnpremSecure {
+		t.Skip("Skipping stateful tests: awscloudtrail_stateful is disabled on-prem")
+	}
 	if strings.HasSuffix(os.Getenv("SYSDIG_SECURE_URL"), "ibm.com") {
 		t.Skip("Skipping stateful tests for IBM Cloud")
 		return
